# perf(logql): Micro-optimizations for IP filter