--- a/pkg/logql/log/ip.go
+++ b/pkg/logql/log/ip.go
@@ -282,7 +282,7 @@ func isHexDigit(r byte) bool {
 // It returns the number of chars in the initial segment of `s`
 // which consist only of chars from `accept`.
 func bytesSpan(s, accept []byte) int {
-	m := make(map[byte]bool)
+	m := make(map[byte]bool, len(accept))
 
 	for _, r := range accept {
 		m[r] = true