**What this PR does / why we need it**:

Small micro optimizations to save a couple allocations for the IP filter in logql.

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/loki/v3/pkg/logql/log
cpu: Apple M2 Pro
                                       │   old.txt   │              new.txt               │
                                       │   sec/op    │    sec/op     vs base              │
_IPFilter/127.0.0.1-10                   5.053µ ± 0%   4.621µ ±  3%  -8.55% (p=0.002 n=6)
_IPFilter/192.168.0.1-192.189.10.12-10   5.080µ ± 0%   4.640µ ±  2%  -8.68% (p=0.002 n=6)
_IPFilter/192.168.4.5/16-10              4.849µ ± 2%   4.378µ ± 27%       ~ (p=0.065 n=6)
geomean                                  4.993µ        4.545µ        -8.98%

                                       │  old.txt   │              new.txt              │
                                       │    B/op    │    B/op     vs base               │
_IPFilter/127.0.0.1-10                   425.0 ± 0%   272.0 ± 0%  -36.00% (p=0.002 n=6)
_IPFilter/192.168.0.1-192.189.10.12-10   425.0 ± 0%   272.0 ± 0%  -36.00% (p=0.002 n=6)
_IPFilter/192.168.4.5/16-10              425.0 ± 0%   272.0 ± 0%  -36.00% (p=0.002 n=6)
geomean                                  425.0        272.0       -36.00%

                                       │   old.txt   │              new.txt              │
                                       │  allocs/op  │ allocs/op   vs base               │
_IPFilter/127.0.0.1-10                   10.000 ± 0%   8.000 ± 0%  -20.00% (p=0.002 n=6)
_IPFilter/192.168.0.1-192.189.10.12-10   10.000 ± 0%   8.000 ± 0%  -20.00% (p=0.002 n=6)
_IPFilter/192.168.4.5/16-10              10.000 ± 0%   8.000 ± 0%  -20.00% (p=0.002 n=6)
geomean                                   10.00        8.000       -20.00%
```

**Which issue(s) this PR fixes**:
Fixes #<issue number>

**Special notes for your reviewer**:

**Checklist**
- [ ] Reviewed the [`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md) guide (**required**)
- [ ] Documentation added
- [ ] Tests updated
- [ ] Title matches the required conventional commits format, see [here](https://www.conventionalcommits.org/en/v1.0.0/)
  - **Note** that Promtail is considered to be feature complete, and future development for logs collection will be in [Grafana Alloy](https://github.com/grafana/alloy). As such, `feat` PRs are unlikely to be accepted unless a case can be made for the feature actually being a bug fix to existing behavior.
- [ ] Changes that require user attention or interaction to upgrade are documented in `docs/sources/setup/upgrade/_index.md`
- [ ] For Helm chart changes bump the Helm chart version in `production/helm/loki/Chart.yaml` and update `production/helm/loki/CHANGELOG.md` and `production/helm/loki/README.md`. [Example PR](https://github.com/grafana/loki/commit/d10549e3ece02120974929894ee333d07755d213)
- [ ] If the change is deprecating or removing a configuration option, update the `deprecated-config.yaml` and `deleted-config.yaml` files respectively in the `tools/deprecated-config-checker` directory. [Example PR](https://github.com/grafana/loki/pull/10840/commits/0d4416a4b03739583349934b96f272fb4f685d15)